Finish moving components from script.js, bug fixes
