+++ src/pages/index.html
@@ -138,6 +138,12 @@
 					</h2>
 				</div>
 			</div>
+
+			<noscript>
+				<link rel="stylesheet" href="noscript.css" />
+				<h2>JavaScript Required</h2>
+				<p>This application requires JavaScript.</p>
+			</noscript>
 		</main>
 
 		<!-- boilerplate popup container -->
