+++ src/styles/base.scss
@@ -17,6 +17,10 @@
 	--canvas-width: 100%;
 	--canvas-height: 750px;
 
+	// styles
+	--box-shadow: 0 0 10px var(--shadow);
+	--radius: 5px;
+
 	// colours
 	--copy: darkgreen;
 	--copy-hover: green;
@@ -33,8 +37,9 @@
 	--shadow: #000;
 	--text: #fff;
 	--link: lightblue;
+	color-scheme: dark;
 
-	&:not([darkmode]) {
+	&[darkmode="false"] {
 		--background-colour: #ddd;
 		--foreground-colour: #eee;
 		--accent: #fff;
@@ -42,11 +47,8 @@
 		--shadow: #ccc;
 		--text: #000;
 		--link: darkblue;
+		color-scheme: light;
 	}
-
-	// styles
-	--box-shadow: 0 0 10px var(--shadow);
-	--radius: 5px;
 }
 
 body {
@@ -93,29 +95,15 @@ main {
 	padding: 1em;
 	position: relative;
 	width: 100%;
-}
 
-button,
-input {
-	background-color: var(--foreground-colour);
-	border: none;
-	border-radius: var(--radius);
-	transition: background-color 0.1s;
-}
+	&[error] {
+		@include mixins.flexbox(column);
+		@include mixins.flexbox-align(center, center);
 
-a,
-button {
-	cursor: pointer;
-	transition: background-color 0.25s;
-
-	&:has(i:only-child) {
-		@include mixins.icon-button();
+		#canvas-container {
+			display: none;
+		}
 	}
-}
-
-a,
-a:active {
-	color: var(--link);
 }
 
 a.link-chip {
@@ -137,6 +125,33 @@ a.link-chip {
 	}
 }
 
+button,
+input {
+	background-color: var(--foreground-colour);
+	border: none;
+	border-radius: var(--radius);
+	transition: background-color 0.1s;
+}
+
+a {
+	color: var(--link);
+}
+
+a,
+button {
+	cursor: pointer;
+	transition: background-color 0.25s;
+}
+
+a:active {
+	color: var(--link);
+}
+
+a:has(i:only-child),
+button:has(i:only-child) {
+	@include mixins.icon-button();
+}
+
 // todo - figure out why this exists
 input[type="text"] {
 	background-color: var(--background-colour);
@@ -155,7 +170,6 @@ input[type="text"] {
 	margin: auto;
 	max-block-size: calc(100svh - 2rem);
 	padding: 1em;
-	// todo - this should probably not be 80ch, test
 	width: min(95vw, 80ch);
 	z-index: 100;
