+++ src/supplement/errors.ts
@@ -1,3 +1,6 @@
+import { Result } from "../index.js";
+import { nullish } from "./index.js";
+
 export class Ok<T> {
 	data: T;
 	ok: true;
@@ -5,6 +8,10 @@ export class Ok<T> {
 	constructor(data: T) {
 		this.data = data;
 		this.ok = true;
+	}
+
+	unwrap() {
+		return this.data;
 	}
 }
 
@@ -18,10 +25,17 @@ export class Err<E> {
 	}
 }
 
-export const AppErrMap = {} as const;
+export function resultConstructor<T>(data: T | undefined) {
+	if (data === undefined) return new Err(undefined);
+	else return new Ok(data);
+}
 
-export type AppErr = keyof typeof AppErrMap;
-
-export function getErr(key: AppErr) {
-	return AppErrMap[key];
+export function isErr(obj: unknown): obj is Err<unknown> {
+	try {
+		const result = obj as Result<unknown, unknown>;
+		if (nullish(result)) return false;
+		return result.ok === false;
+	} catch (e) {
+		return false;
+	}
 }
